pyproject.toml
[tool.poetry]
name = "ujumbe"
version = "1.0.0"
description = ""
authors = ["BrianLusina <12752833+BrianLusina@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.92.0"
uvicorn = {extras = ["stadard"], version = "^0.20.0"}
celery = {extras = ["redis"], version = "^5.2.7"}
python-dotenv = "^1.0.0"
loguru = "^0.6.0"
httpx = "^0.23.3"
confluent-kafka = "^2.0.2"
phonenumbers = "^8.13.6"
nanoid = "^2.0.0"
dependency-injector = "^4.41.0"
sqlalchemy = "^2.0.7"
alembic = "^1.10.2"
inflection = "^0.5.1"
twilio = "^7.17.0"
requests = "^2.28.2"
attrs = "^22.2.0"
psycopg2 = "^2.9.6"

[tool.poetry.group.dev.dependencies]
black = "^23.1.0"
flake8 = "^6.0.0"
isort = "^5.12.0"
mypy = "^1.0.1"
pylint = "^2.16.2"
coverage = "^7.2.1"
pre-commit = "^3.1.1"
pytest-cov = "^4.0.0"
pytest = "^7.2.1"
locust = "^2.15.0"
faker = "^17.5.0"
testcontainers-postgres = "^0.0.1rc1"
testcontainers-kafka = "^0.0.1rc1"

# Pytest Options
[tool.pytest.ini_options]
addopts = [
    "--import-mode=importlib",
]
pythonpath = "app"
norecursedirs = "tmp"
testpaths = "tests"
filterwarnings = [ "ignore::DeprecationWarning" ]
markers = [
    "integration: integration test",
    "unit: unit test",
    "e2e: end to end test"
]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
